GR
reorganizacion
Se creo un playerpathmanager para que se encargue de las cosas que solo el del player puede hacer
Se modifico el layer playerpath a solo path ya que todos los autos lo usaran
Se modifica los prefabs para que tengan sus layers correspondientes
