+++ Assets/Prefabs/Gameplay.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 1a80b343f1d34574fa38f206d06b2b9f
+guid: d787a44f71b7a364387b537387bcdeb9
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
