+++ Assets/Prefabs/Gameplay/Car Obstacle Variant.prefab.meta
