+++ Assets/Prefabs/Gameplay/Enemy.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 1a80b343f1d34574fa38f206d06b2b9f
-DefaultImporter:
+guid: ee48f842e3dfbaa429f9194d9c1a1b00
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
