--- Assets/Scripts/Editor Scripts.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6416a8e25c544928b82a9d64673dc98e
-timeCreated: 1682213108
