+++ Assets/Scripts/Gameplay/Car/Path/PlayerPathManager.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 2537e73eb0f44c24bb9fccc5827995fa
+timeCreated: 1682283990
