+++ Assets/Scripts/UI/MenuPrincipalUI.cs
@@ -9,5 +9,13 @@ namespace ProyectM2.UI
         {
             SceneManager.Instance.ChangeScene(new Scene("Level "+level, Scene.Type.Gameplay));
         }
+#if UNITY_EDITOR
+        [SerializeField] private string debuScene;
+        [ContextMenu("Play Debug")]
+        public void Play()
+        {
+            SceneManager.Instance.ChangeScene(new Scene(debuScene, Scene.Type.Gameplay));
+        }
+#endif
     }
 }
