+++ ProjectSettings/EditorBuildSettings.asset
@@ -17,6 +17,9 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Gameplay.unity
     guid: 129f2df9338d2f94cab40d825b2db7f3
+  - enabled: 1
+    path: Assets/Scenes/Nuestras Escenas (no meter en build)/Gabriel.unity
+    guid: 1a80b343f1d34574fa38f206d06b2b9f
   m_configObjects:
     com.unity.adaptiveperformance.loader_settings: {fileID: 11400000, guid: fba86f12aa250474fbc25ffcbcf0b7c3,
       type: 2}
